fix: resolved comments
