+++ package.json
@@ -106,7 +106,6 @@
   },
   "dependencies": {
     "postcss": "^8.3.6",
-    "react-infinite-scroll-component": "^6.1.0",
     "react-table": "^7.7.0",
     "react-table-sticky": "^1.1.3",
     "react-virtualized-auto-sizer": "^1.0.6",
